feat(cli): validate Directus version
