+++ packages/cli/src/lib/services/migration-client.ts
@@ -8,12 +8,15 @@ import {
   rest,
   RestClient,
   RestConfig,
+  serverInfo,
 } from '@directus/sdk';
 import { Inject, Service } from 'typedi';
 import pino from 'pino';
 import { LOGGER } from '../constants';
 import { ConfigService, isDirectusConfigWithToken } from './config';
 import { getChildLogger } from '../helpers';
+import { compareVersions } from 'compare-versions';
+import { Cacheable } from 'typescript-cacheable';
 
 @Service()
 export class MigrationClient {
@@ -78,5 +81,53 @@ export class MigrationClient {
 
     await client.setToken(token);
     return client;
+  }
+
+  /**
+   * Get the server version
+   */
+  @Cacheable()
+  protected async getDirectusVersion() {
+    const client = await this.get();
+    const info = await client.request<{ version?: string }>(serverInfo());
+    if (!info?.version) {
+      throw new Error('Could not get the Directus version');
+    }
+    return info.version;
+  }
+
+  /**
+   * This method compares the Directus instance version with the given one.
+   */
+  protected async compareWithDirectusVersion(
+    version: string,
+  ): Promise<'equal' | 'greater' | 'smaller'> {
+    const directusVersion = await this.getDirectusVersion();
+    const diff = compareVersions(version, directusVersion);
+    if (diff === 0) {
+      return 'equal';
+    } else if (diff > 0) {
+      return 'greater';
+    } else {
+      return 'smaller';
+    }
+  }
+
+  /**
+   * This method validate that the Directus instance version is compatible with the current CLI version.
+   */
+  async validateDirectusVersion() {
+    const directusVersion = await this.getDirectusVersion();
+    if ((await this.compareWithDirectusVersion('10.0.0')) === 'greater') {
+      throw new Error(
+        `This CLI is not compatible with Directus ${directusVersion}. Please upgrade Directus to 10.0.0 (or higher).`,
+      );
+    }
+    if ((await this.compareWithDirectusVersion('11.0.0')) === 'greater') {
+      throw new Error(
+        `This CLI is not compatible with Directus ${directusVersion}. Please use \`npx directus-sync@2.2.0 [command]\` or upgrade Directus to 11.0.0 (or higher).`,
+      );
+    }
+    this.logger.debug(`Directus ${directusVersion} is compatible`);
   }
 }
